feat: implementasi fitur read untuk Product dengan menampilkan data relasi dari Category beserta fitur searching, pagination, dan filter

+++ database/migrations/2026_06_02_030118_create_products_table.php
@@ -19,10 +19,11 @@ return new class extends Migration
           ->onDelete('cascade');
 
     $table->string('nama_produk');
-    $table->decimal('harga', 12, 2);
+    $table->integer('harga');
     $table->integer('stok');
     $table->string('merk');
     $table->text('deskripsi');
+
     $table->timestamps();
 });
     }
